requirements.txt
faster-whisper
funasr
modelscope
torch
torchaudio
langchain
langchain-community
langchain-openai
langchain-ollama
langchain-huggingface
sentence-transformers
faiss-cpu
pydub
python-dotenv
requests
dashscope
zhipuai
numpy<2.0.0
pyinstaller
pyaudio
webrtcvad
sounddevice
streamlit
streamlit-mic-recorder
